Adds tracking of ambiguous range alignments to the Chapel range type.
ChapelRange.chpl was renamed as ChapelRangeBase.chpl, and the range type was
renamed as rangeBase.  The range type was then reincarnated as a wrapper for
rangeBase, adding tracking of an _aligned flag, which allows determination of
when the alignment of a range is ambiguous.

These tests were updated because range expressions that attempt to iterate
ambiguous ranges are now illegal:
       test/release/examples/primers/ranges.chpl
       test/release/examples/primers/ranges.good
       test/types/range/diten/rangeCountBy.chpl
       test/types/range/diten/rangeCountBy.good

A number of .chpl files simply required renaming _low, _high and _stride as low,
high and stride (without the leading underscores), because the former are now
hidden within the rangeBase record.
       test/arrays/marybeth/CMO_array.chpl
       test/ranges/bradc/bigStride.chpl
       test/studies/hpcc/common/bradc/hpccMultilocale.chpl
       modules/layouts/GPUDist.chpl
       modules/dists/CyclicDist.chpl
       modules/dists/BlockCycDist.chpl
       modules/dists/BlockDist.chpl
       modules/internal/ChapelArray.chpl
       modules/internal/DefaultRectangular.chpl
       modules/internal/DefaultAssociative.chpl

The results of this test changed because the order of inclusion of the standard
modules changed.
       test/modules/bradc/printModStuff/foo.good

This was rewritten because with Jonathan's recent changes the return type of
cells.dim(d) became ambiguous.  This case is not yet working.
       test/studies/amr/lib/grid/Grid_def.chpl

I added some more detail to the error messages.
       test/domains/marybeth/test_compare_range.good
       test/classes/hilde/explainInstantiation.good
       modules/internal/ChapelBase.chpl

The output format of a range type has changed:
       test/types/range/claridge/strided_range_constructor.chpl

I added a comment describing the meaning and usage of FLAG_RANGE.
       compiler/include/flags_list.h

Added snide comment on how the isBoundedIterator predicate is used.
       compiler/resolution/lowerIterators.cpp


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18549 3a8e244f-b0f2-452b-bcba-4c88e055c3ca