Fixing a bug in Block-Cyclic
Block-Cyclic had a bug that almost never showed up (until I checked
in ptrans to run on three locales) in which I would shift an index
value to 0-based coordinates and then drop it on the floor using
the original index value instead.  Oops.  This commit fixes the bug
and files a case that shows it (in a simpler context than ptrans)
for future reference (because ptrans is likely to keep evolving).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18554 3a8e244f-b0f2-452b-bcba-4c88e055c3ca