Moving multirealm-related stuff out of the release
Since multirealm is most useful/sensible under PVM and we're not
planning on releasing PVM support in the 1.3 release, I moved the
multirealm-related documentation and testing out of the release and
into other places.  In particular, I moved README.multirealm into a
new doc/archive directory and moved the multirealm.chpl test in the
release into the normal test/multirealm directory.  Updated the
related READMEs to reflect the change.  Note that this doesn't remove
the code/support for multirealm, it merely avoids promoting it as
something to try.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18569 3a8e244f-b0f2-452b-bcba-4c88e055c3ca