Adding the variants/ directory to hpcc/
Adding the variants/ directory we discussed to the hpcc/ directory
to capture more elegant (or sufficiently interesting) variants on
our official entries (which sometimes sacrifice elegance in order
to get better performance with today's compiler).

At this point, I added two variants, one for promoted stream, one
for RA using a much cleaner loop.

I inherited the .perf* files from the originals, but didn't copy
the .graph files, figuring we'd want to overlay the performances
together on a single graph.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18577 3a8e244f-b0f2-452b-bcba-4c88e055c3ca