Pass the -hfp1 flag to the Cray compiler when Chapel's --no-ieee-float flag is specified.
When using the Random module, -hfp2 (the default) produces numbers
that do not match the expected values, possibly producing
non-pseudorandom streams.  I have not yet figured out if this is a
problem with CCE.  It could also be assumptions we make about signed
integer overflow.

BTW, the craycc man page does not explicitly state that -hfp0 conforms
100% to the IEEE floating-point standard (as is implied by Chapel's
--ieee-float).

This fixes the following XT CCE regressions:

     [Error matching program output for performance/sungeun/assign]
     [Error matching program output for performance/sungeun/dgemm]




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18588 3a8e244f-b0f2-452b-bcba-4c88e055c3ca