Adding a new updated version of the locales.chpl test.
This update adds a lot more hand-holding in the form of
comments, adds a lot of discussion about the Locales
array, here, queries on locale types, some discussion
of PGAS, orthogonality of locality and parallelism, 
data-driven on-clauses, etc.  I ended up keeping the
distributed memory linked list example (but with more
motivation for why it is interesting) but dropped the
distributed memory fibonacci example that was more
confusing/artificial than useful to me.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18599 3a8e244f-b0f2-452b-bcba-4c88e055c3ca