Updating future to reflect current status
Sung thought we could retire this future now that config params are
simply parsed into the code.  To try and capture the original intent
of the test, I declared the type of the config param and tried to
initialize it on the command line with an illegal value.
Unfortunately, the resulting error message seems incorrect.  Updated
the .future to reflect this.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18619 3a8e244f-b0f2-452b-bcba-4c88e055c3ca