I had mistakenly removed all mention of record inheritance (or moved it to future sections).
I discovered that that is not quite correct.  Currently, field inheritance in records *is* supported.

So, I split the two concepts and reinstated the recordInheritance example.

The rest of the files contain general cleanup.  I made a pass through the entire spec, looking for REVIEW and TODO comments with my name on them, and followed those instructions if the changes seemed relatively minimal.  Some had already been resolved, so all I had to do was remove the comment.

Also, because I'm fussy and I already had the files open, I moved some text around to get all of the overfull hbox warnings to report a deviation of 25 points or less.  There are 72 points to the inch, so this means we are accepting some instances of text jutting into the right margin by about 1/3 of an inch.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18631 3a8e244f-b0f2-452b-bcba-4c88e055c3ca