Making a couple of improvements to our routine to chunk up work
This commit makes a couple of important improvements in the routine
we use to chunk up work across nodes/tasks.  It is probably still
not perfect, but is definitely an improvement on the previous draft:

1) It uses an improved implementation of ceil(x/y) where x and y are
   integers that is less susceptible to wraparound

2) (probably more importantly) It performs some initial math using
   uint(64)s to minimize the chances of wraparound.

More thought should probably be given to whether wraparound in #2
is still possible/likely to occur for realistic cases that we'll
see, and also whether such wraparound would be benign (which I
think I had always assumed for the previous version that used
the range/domain index type -- erroneously as it turns out) or not.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18647 3a8e244f-b0f2-452b-bcba-4c88e055c3ca