Took a pass over the '_extern' technical note
The original motivation was to add a description of extern consts and
extern classes and assignment semantics for extern records, and I did
all these things, but then I also took a pass over the entire file to
try and arrange the information into a more logical sequence.  (The
previous sequence more or less reflected the order in which the
features were implemented which tended to require reading the whole
thing).

I also added FAQs about pointer-to-type arguments and passing arrays
out to external functions since we get asked both of those questions a
lot.  In both cases this was marked as a temporary way of doing
things, not the ultimate solution.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18665 3a8e244f-b0f2-452b-bcba-4c88e055c3ca