Update of primers README and files.
Added new test names to the README and cleaned up the descriptions a
bit.

Changed header comments in the primers to be a little more consistent
and added the name of the primer as the first line of the comment.
Some of the primers still refered to themselves as examples since we
used to call them examples until a couple of releases ago.

Replaced the word 'function' with 'procedure' where ever it seemed
appropriate.  This includes a one word change in the output of
slices.chpl (function=>procedure).  Sorry for that change, but it
seems simple and necessary.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18667 3a8e244f-b0f2-452b-bcba-4c88e055c3ca