Added a README for volatile types; minor wording change to README.local
I had planned to drop volatile types into the spec, but it's not as
minor a change as I had hoped due to the syntax diagrams (we support
most scalar types as volatiles but not all (strings, complex), so I'm
punting on this for this release.  The technical notes seem like a
nice place to incubate them for the time being anyway.

Made a minor wording change to README.local to try and remove the
implication that local should be avoided at all costs.  :)

Added README.volatile to the doc/README directory listing and
re-alphabetized the list.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18677 3a8e244f-b0f2-452b-bcba-4c88e055c3ca