Reverted all changes to the spec and README.firstClassFns from r18653 to r18682, in preparation for reverting just r18653.
Rationale:

In today's Chapel standup meeting we decided that we did not like the new

  procedures = (non-iterator procedures) + iterators

terminology. So for the time being we decided to revert back to

  functions = procedures + iterators

This is the easiest way to handle the lack of a clear leader in terminology,
given very little time before the docs freeze for the 1.3 release.

This reversal presents a much smaller terminology change to the end users
for this release. We can still consider/implement sweeping changes later
(like functions->procedures or others).

I left alone the function->procedure changes in r18667 (primers) for expediency
and because they are still valid (in most, if not all, cases, the new uses
of "procedure" refer to non-iterator procedures).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18683 3a8e244f-b0f2-452b-bcba-4c88e055c3ca