Updating future with additional info.
From the future:

This fails for --no-local and CHPL_COMM!=none builds because the
reference type is for uint(8) is deleted.  This happens because the
cullOverReferences pass deems it not necessary, and since this pass
runs before parallel, as far as it is concerned, it isn't.

The cullOverReferences pass could be more conservative, but then it
would capture far more references than needed.  Another option would
be to move the pass after parallel (and handle the cases introduced in
parallel), but I really have no idea what implications that has.

We should probably have assertions in codegen for the case where the
ref type is not found.

As a workaround, if you pass the array in question to another
function, say writeln(), the ref type is not deleted.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18706 3a8e244f-b0f2-452b-bcba-4c88e055c3ca