Changed Stats dir for cyclic tests.
Changed CHPL_NIGHTLY_STATDIR to be
/data/sea/cascade/NightlyDists/Stats.cyclic to avoid possibly conflict
with Block dist.  I think the stats stuff is not quite working
properly with respect to naming (the dist names seem to be dropped).

In talking with Brad, some stuff broke when testing got moved over to
chapelu, so I'm leaving the task of getting it all working again for
when we revamp the stat web pages.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18711 3a8e244f-b0f2-452b-bcba-4c88e055c3ca