Range count bug fix -- allow and use tighter bounds.
The range specification 10..100 by 10 # 10 was generating a run-time error since
the bounds desired by the algorithm (10..109) lay beyond the user-specified
bounds.  Consensus so far is that the bounds returned by # should be clipped by
the user's bounds without warning.

We already perform a check to ensure that the original range is at least large
enough to produce the number of indices requested, so the only cases where
clipping will apply are those in which the specified opposite bound of the range
is between 0 (a tight bound) and the stride (non-inclusive on both ends).  


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18714 3a8e244f-b0f2-452b-bcba-4c88e055c3ca