Fix launcher bugs that showed up when using the Jaguar system.
Srini happened to be using the new launcher today and found that it
didn't work on the Jaguar system.  There were 2 problems with
chpl_run_utility1K() which runs another (short) program.  First, I
forgot to create a pipe for and dup stderr (20 points for Brad for
guessing that without even looking at the code!), and then I was
assuming that reading 0 bytes meant that the program had exited.

I also added a timeout argument to the select() calls since a NULL
timeout can cause select() to block indefinitely.

NOTE: Currently, if a program never exits *and* writes less than 1K
characters to stdout and stderr combined, this function will never
exit.  This is not a recommended us of the function, but I should
eventually come up with a time out strategy.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18719 3a8e244f-b0f2-452b-bcba-4c88e055c3ca