More revamping of the way we use FUTURES and STATUS.devel.
Add a new script (futures_stats) that takes a STATUS.devel and FUTURES
file and print the info about futures that are in one file and not the
other.  The script assumes a particular format for both files.  Add a
make target 'future_stats' that generates the FUTURES file and runs
the script.

Delete cat_futures and list_futures because they aren't really useful
anymore.

  cat_future: Who wants to cat all the futures into a single
    file?  You can do it with a single command if you do.

  list_futures: This assumed we assigned futures to people which we
    don't any more, and if we started again, we'd probably use a
    different format.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18725 3a8e244f-b0f2-452b-bcba-4c88e055c3ca