Fixed the much needed lack of != on ranges and a typo in == and != on domains. Also clarified an error message. Added a test for those (but more test code is surely wanting).
Motivation: the following used to be broken:

  (0..(1:int)) != (0..(1:uint)) // true
  [0..(1:int)] == [0..(1:uint)] // false
  [0..(1:int)] != [0..(1:uint)] // true

Interesting: I do not quite know why it worked correctly when
the two ranges were of the exact same type, e.g.:
  (0..1) == (0..1)
  [0..1] == [0..1]
  and the above with !=

It probably goes like this: in the absence of a user-defined != on ranges,
the compiler created one - because 'range' is a Chapel record.
That default one compares *all* fields, including the type field idxType.
So there.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18726 3a8e244f-b0f2-452b-bcba-4c88e055c3ca