Updated range.== to allow ambiguous ranges to be equal if their representations match.
M    modules/internal/ChapelRange.chpl
Equivalence of two ambiguously-aligned ranges added as a special case.
The case is coded this way, because the base-class implementation of ==
compares the start and end fields rather than the low and high.  Thus, the
general case could allow ambiguously aligned ranges with different bounds to
compare equal.  One would probably have to use slicing to produce such contrived
ambiguous ranges.  The simple (..1 by 2) vs. (..0 by 2) case would not slip
through.

M    modules/internal/ChapelRangeBase.chpl
Updated an error message to better match the test.
Fixed an obscure bug in the count operator which allows counting off more
elements than lie within the bounds, if the count is negative.
This is the problem that originally set me on the trail of the 10..100 by 10 #
10 bug, and now I have finally gotten back to it (and fixed it).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18739 3a8e244f-b0f2-452b-bcba-4c88e055c3ca