Fix bug in test_module_Norm.chpl found by using CCE.
This fix should get rid of the regression for CCE.

This test saves results of various Norm operations and then compares
the results from the same computation with modified arrays (e.g.,
sliced, aliased, rank-changed, etc.).  One of the comparisons was
looking at the wrong saved (integer) value.

This is actually a floating point precision issue.  The saved values
are scaled and then converted to integers to avoid noise in the
comparison.  CCE 32- vs. 64-bit floating point results for the Norm
computations seem to differ a bit more than for the other compilers
(i.e., for the other compilers, 32- and 64-bit scaled integer results
were the same), which is why this regression only showed up for CCE.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18750 3a8e244f-b0f2-452b-bcba-4c88e055c3ca