Normalized the code a bit and added comments.
I made sure that each static function was mentioned, in order, in a block of forward static function declarations at the beginning of the file (serving as a kind of index).

Moved some functions which had been added later down below the main routine: normalize(), so that the organization is a bit more top-down. (I only did this at the top level.)

Added some comments to the function change_method_to_consructor() in response to Aaron Stacy's bug report.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18760 3a8e244f-b0f2-452b-bcba-4c88e055c3ca