Added test cases for duplicate auto destroys.
These are obtained by a reduction of a version of KM's graph500 code.
The common tune is an array of records, with each element
explicitly initialized via an assignment from a 'new' on the record type.

array-of-records-crash-1.chpl and array-of-records-crash-2.chpl
crash upon triplicate TILDA on a sync variable.

array-of-records-crash-3.chpl crashes upon a duplicate TILDA
on a domain.

Interestingly, uncommenting the explicit type of the array in the tests
eliminates the symptoms. (I have not investigated this.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18789 3a8e244f-b0f2-452b-bcba-4c88e055c3ca