Added a version of Kristi Maschhoff's Graph500 code provided to us at the end of March '2011.
This code is based on the XMT version of the Graph5000 benchmark submitted by
John Feo <john.feo@pnl.gov> and Kristi Maschhoff <kristyn@cray.com> in 2010.

NUMLOCALES=4 was chosen arbitrarily.

Changes from Kristi's version:

* smaller problem size: SCALE 8->6, EDGEFACTOR 16->8

* output_result() simply prints "done"
  vs. calling an external function

---

NOTE: This program crashes under GASNet with
  *** Caught a fatal signal: SIGSEGV(11) on node XX/YY

Therefore I am adding it to REGRESSIONS right away.

The crash is intermittent, but I often get it when running the program
right after compilation.

For the record, when running with GASNET_BACKTRACE=1,
the thread that catches SIGSEGV has the following at the bottom of the stack:

#22 0x00000000006bc483 in gasnet_barrier_wait (id=0, flags=1) at /users/vass/chapel/third-party/gasnet/GASNet-1.14.2/extended-ref/gasnet_extended_refbarrier.c:1279
#23 0x0000000000690800 in chpl_comm_barrier (msg=0x8450b9 "chpl_comm_exit_all") at comm-gasnet.c:394
#24 0x000000000066e5fd in chpl_exit_common (status=0, all=1) at chplexit.c:20
#25 0x000000000066e647 in chpl_exit_all (status=0) at chplexit.c:34
#26 0x000000000066a071 in main (argc=3, argv=0x7fff411f5d68) at main.c:104


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18790 3a8e244f-b0f2-452b-bcba-4c88e055c3ca