Updated graph500/v2 to a mid-May version of Kristi Maschhoff's Graph500 code.
It's mostly the same as 'v1' (see r18790).
Notes:

* DISTRIBUTION_TYPE is changed to "" from "BLOCK".
(With "BLOCK", one can get a different symptom for the crash mentioned next.)

* If the explicit type of the Vertices field in defs.chpl
is commented out, this version exhibits the crashes
exemplified by test/users/vass/km/array-of-records-crash-*.

* Kristi's version changes 'forall' to 'for' at the top of constructGraph.chpl
with the "Note that code currently crashes if this is run in parallel".
I did not observe those crashes so switched that 'for' back to 'forall'.

* The "out of bounds" error is expected - it's a snapshot
of Kristi's development.

---

More notes on r18790  (meant to be part of r18790 commit message):

* The computation is nondeterministic even though a fixed seed
is provided to the random number generator.
This could be due in part to the permutation code at the end of
edge construction.

Even earlier in the program, min/max vertex IDs change
(e.g. min = 1 (typical) or 2, max = 255 or 256 (typical)).
Neither I nor Kristi know why. Maybe the random number generator
is still nondeterministic?

* Therefore I added the flag ENABLE_PRINTOUTS to guard writeln
that print nondeterministic info (including timings).

* While I mention in r18790 intermittent gasnet crashes,
I have not observed 'v1' to crash under comm=none or --no-local.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18792 3a8e244f-b0f2-452b-bcba-4c88e055c3ca