Removed a questionable pragma-marked assignment from chpl__initCopy(). Also removed a .future for a test that should be working now.
The line:

    pragma "no copy" pragma "insert auto destroy" var ee = e;

with the subsequent references to 'e' replaced by 'ee' was introduced
by Steve in r16033 in an effort to fix memory leaks.

However, this line introduces one autodestroy too many in
  test/users/vass/km/array-of-records-crash-3.chpl
(and probably in array-of-records-crash-1,2 as well),
resulting in either double free() of a pointer
(if it's a record with a sync var field)
or a double decrement of a reference counter below 0
(if it's a record with a domain field).

Following Brad's suggestion, I removed that line and replaced
the reference to 'ee' back to 'e'.

This fixes the crash in array-of-records-crash-3.chpl, while passing
the standard comm=none and comm=gasnet parallel testing and leaking
the same amount of memory as the baseline in
  studies/madness/aniruddha/madchap/recordBug/test_likepy
which is the test that Steve targeted in this commit.

So we'll give it a go, even though we do not have
a solid explanation of what the issue was and why it's fixed now.

BTW there were also a couple of lines
  leaderIndexCopy->addFlag(FLAG_INSERT_AUTO_DESTROY);
added to the compiler in that same commit.
Makes me wonder whether those could/should go as well...


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18793 3a8e244f-b0f2-452b-bcba-4c88e055c3ca