Updated info on this future.
Looked a bit into this future to see if it was due to the same record
constructor/destructor woes that I have been working on.  Turns out
that the answer is no, but it is related to our reference counting
scheme which I am preparing to revamp, so I propose we let this be for
now.

From the future:

     bug: 'data' field in DefaultRectangularArr freed when still in use

     The data field (class _ddata) is being freed in dsiReallocate() via a
     direct call to dsiDestroyData(), but it may still be in use due to
     calls that copy the field (i.e., reindexing, slicing, or rank change).

     There needs to be some sort of tracking of references to the data
     field.  It seems possible to roll this into the current reference
     counting on arrays.

     This tests periodically seg faults with --no-local and GASNet.  It
     consistently fails valgrind testing with --no-local.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18805 3a8e244f-b0f2-452b-bcba-4c88e055c3ca