Clean up Swap benchmark.
Removed unnecessary test cases.

Moved code that optimized for data-locality to its own file. 
For example, swap.chpl and swapOn.chpl use atomic transactions. 
However, the latter additionally uses the on statement to optimize 
for the data locality of the second array element. For each of these
cases, the lock based kernels additionally defines a config param 
(useAffinity) that optimizes for the locality of locks as well.




git-svn-id: http://svn.code.sf.net/p/chapel/code/branches/collaborations/gtm@18813 3a8e244f-b0f2-452b-bcba-4c88e055c3ca