Added examples that were missing in Statments.tex as noted by Mark Potts.
Also did a general scan of "chapeloutput" sections to see if any additional ones should be converted to "chapelprintoutput".

I skipped chapeloutput sections that were preceded by "chapelnoprint" or "chapelpost" sections.  The decided whether to print the remaining "chapeloutput" sections based on context.

Where the situation was ambiguous, I erred on the side of printing vs not.  When red-lining the spec, it is easier to make unwanted to stuff disappear than to divine what is missing.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18814 3a8e244f-b0f2-452b-bcba-4c88e055c3ca