Initial implementation for two versions of the Bailey, Borwein, Plouffe algorithm.
These print out the first 10000 hex digits of Pi using the Plouffe formula and only integer arithmetic.

increm.chpl -- is memory intensive.  It stores the remainders from the previous step so that calculation of the next step is incremental.
nomem32.chpl -- is computation intensive.  It calculates the place-value contribution of each fraction ab initio on each iteration.  But it performs the computation in chunks of 28 bits, so it effectively computes 7 hex digits in parallel (SIMD).

Both algorithms are "embarrassingly parallelizable", which is a next step.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18833 3a8e244f-b0f2-452b-bcba-4c88e055c3ca