Adding some simple string tests and a couple of futures.
I ran into a bug with calling writeln() on a class with a record with
a string.  I looked at the generated code and everything leading up to
the writeln() looks correct.  The result in some cases is highly
unexpected.

I also wrote a few other simple tests that are related.  They all
pass, and I've checked them in because the generated code for the
correct tests might be helpful when debugging.

The string2*.chpl tests directly pass a the data being tested directly
to writeln(), and the string1*.chpl tests wrap that data in a tuple
with a string.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18855 3a8e244f-b0f2-452b-bcba-4c88e055c3ca