Updating REGRESSIONS based on past week or so's runs
Mostly just nondeterministic timeout noise

(Sung, note that one test seems to be failing due to a race in the
test itself -- an associative domain is checking for index membership
and I think can get one of two answers (if not more) in a race-y way.
See first entry in svn diff for this commit).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18856 3a8e244f-b0f2-452b-bcba-4c88e055c3ca