Making -valgrind + CHPL_COMM!=none valgrind child processes
I just noticed that when we use start_test with -valgrind on, we
don't valgrind the child processes of a GASNet execution, which
can be confusing if you happen to be working with GASNet.  Changed
this so that if we're testing with valgrind and CHPL_COMM != none,
we valgrind the child processes.

Hmm, should that really have been if CHPL_LAUNCHER != none...?




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18859 3a8e244f-b0f2-452b-bcba-4c88e055c3ca