Generate graphs for memleaks testing.
This is a sort of temporary kludge until we figure out how we're going
to divide up the graphs to multiple web pages.

The nightly script will update the memleaks data file when doing
memleaks testing, and start_test will generate the graph when doing
performance testing.  Since performance testing and memleaks testing
are run separately, this may result in a one day lag.

Currently, a soft link to the memleaks data file exists in each
performance directory (brad-lnx and chap03), so the exact same graphs
will appear in the results from each of these tests.

I've put a README in NightlyPerformance and NightlyMemLeaks
summarizing the current situation.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18861 3a8e244f-b0f2-452b-bcba-4c88e055c3ca