Future against declared array-of-array formal types
None of the ways that one might logically want to declare a formal
argument to have array-of-array types works today, unfortunately.
Some are syntax errors (!) others result in a problem deeper in the
compiler, apparently due to the interpretation of the type as a forall
expression.  The workaround for today is to have the type be
completely inferred (checking in a test for this as
arrayOfArrayArg-inferred.chpl).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18871 3a8e244f-b0f2-452b-bcba-4c88e055c3ca