Reclassified future as unimplemented feature and added to STATUS
I dug up the "inability to return an array from a recursive procedure"
future that Kyle ran into the other week and reclassified it as an
unimplemented feature because I think that's really more accurate.
That is, there really ought to be a way to return arrays from a
recursive procedure even if we haven't been very clear about what the
limitations there are.  Added to STATUS.devel since I was surprised
not to see it there to begin with (not surprising, since it was filed
as a semantic future).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18873 3a8e244f-b0f2-452b-bcba-4c88e055c3ca