Adding # operator for dense rectangular domains and arrays
This commit adds an implementation of the # operator for dense
rectangular domains and arrays, defined to take a tuple with a # of
elements matching the domain/array's rank, or a standalone integer in
the case of 1D domains/arrays.  The implementation of the domain
applies the tuple values to the component ranges and uses the
resulting ranges to slice the domain.  The array applies the tuple to
the array's domain and then uses the result to slice the array.

Updated the spec to include this capability and added a few tests,
including a cleaner version of the code to split a domain in half that
I sent Kyle the other week.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18875 3a8e244f-b0f2-452b-bcba-4c88e055c3ca