Disabled promotion of the '..' operator
This commit disables promotion of the '..' operator.  Previously,
an expression like 1..3..5 was interpreted as (1..3)..5 which was
interpreted as promoting the low bound of the "..5" operator with
the range 1..3, yielding 1..5, 2..5, 3..5.  While cute, this was
more confusing than anything, and not particularly useful.  This
commit disables this form by overloading the build_range operator
(that's used to implement "..") to catch any overloads that aren't
already specified and warning that the bounds must be integers of
compatible type.  Added a few new tests of this capability and
updated the .good files for two tests that exercised this in a
few guises (neither of which was useful -- one was a Steve puzzle
for fun; the other locked in an error case for ".. ..").



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18876 3a8e244f-b0f2-452b-bcba-4c88e055c3ca