Adding some sync/single tests (and deleting a future I didn't mean to add in r18866)
I wanted to capture some additional bad behavior with regard to
sync/single in records.  I hesitated at first and didn't check these
in with r18866 because the tests that expose incorrect behavior do so
when argument intents are specified and since we're almost nearly
close to on the cusp of redefining argument intents and the defaults,
I was thinking about waiting.  But with today's mail regarding single
vars in records, I thought it would be worth it to document any bad
behavior that exists today.

Also, I'm deleting a future I accidentally checked in for
types/records/sungeun/arrayInRecordInClass.future in r18866.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18877 3a8e244f-b0f2-452b-bcba-4c88e055c3ca