Added a couple of tests that crash the compiler, arising from my DimensionalDist work.
I did not investigate.
The crashes seem to have started after I added 'seena' and related code.

It is quite likely that these tests can be trimmed down to very small sizes.

I did not record the expected *execution* output to avoid extra work -
instead I will do that for the versions of this code that actually *compile*.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18878 3a8e244f-b0f2-452b-bcba-4c88e055c3ca