Fixing result widths of complex operations
Most of our built-in math operations that generated complex values
were returning default-width complexes (which is to say 128 bits)
regardless of the input width (e.g., adding 64-bit complexes should
generate a 64-bit complex to be consistent with our other floating
point operations).  Fixed this by declaring an explicit return width
as a function of the input widths and added a test to lock in this
behavior.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18884 3a8e244f-b0f2-452b-bcba-4c88e055c3ca