This test has a race that has never been debugged, so making it a .notest
This test has always had a race that we have never debugged, and it
isn't our primary chameneos implementation, so I'm having the test
system skip it for the time being.  We can restore it if/when someone
determines how to resolve the race and decides that it's worth
testing.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18886 3a8e244f-b0f2-452b-bcba-4c88e055c3ca