Fixing bug pointed out by Jens Breitbart
Jens sent in a program that ran into an internal error due to
declaring a var function that did not return anything.  I
inserted a compiler semantic check error against this case
and also improved the internal error to print the location
of the call that caused the error to help provide some clue
as to where it's going (if others run into it along some
other path down the road).

Moved the future that I originally filed against this into
a subdirectory where it fits better and added a second
case (a return with no expression being returned) that
caused similar problems.

Had to update a few tests that had this behavior in code
that wasn't called or code that was testing compilerError
behavior.  In one case, I just made the function non-var;
in the other, I made the function return something to
preserve the original behavior that it was testing.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18914 3a8e244f-b0f2-452b-bcba-4c88e055c3ca