Adding a new primer to the release on leader/follower iterators
In working up an answer to Jens' email about leader/follower iterators
today, I decided that a primer on this topic would be in order, and
also would serve as a good ramp-up for the leader/follower iterator
paper I'm hoping to write for PGAS in the next few weeks.  I pulled
the leader/follower content from iterators.chpl and put a reference to
this new test instead.  Comments/feedback appreciated!



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18915 3a8e244f-b0f2-452b-bcba-4c88e055c3ca