Change level sets to be associative arrays.
Changed the level sets to be associative arrays rather than sparse so
that we could remove the explicit locking from the kernel 4 code.
This also required that we change the local sum for validation into a
sync since it is no longer protected by the lock.  We do not enable
validation when timing so it should not affect the performance
results.

Kristi and I agreed to do this for now to clean up the kernel 4 code,
though there may be a time in the future when we change this to
something else (possibly 1-d array).

It will be interesting to see if there are performance implications.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18920 3a8e244f-b0f2-452b-bcba-4c88e055c3ca